web change

+++ controllers/CartController.php
@@ -106,9 +106,9 @@ class CartController extends AppController {
             $order_items->order_id = $order_id;
             $order_items->product_id = $id;
             $order_items->product_name = $item['name'];
-            $order_items->price = (int) $item['price'];
+            $order_items->price = (float) $item['price'];
             $order_items->qty_item = $item['qty'];
-            $order_items->sum_item = $item['qty'] * (int) $item['price'];
+            $order_items->sum_item = $item['qty'] * (float) $item['price'];
             $order_items->save();
         }
     }
